requirements.txt
aniso8601==7.0.0
appdirs==1.4.4
appnope==0.1.2
asgiref==3.5.2
audioread==2.1.9
backcall==0.2.0
branca==0.4.2
build==0.10.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.7
citric==0.0.5
click==8.1.3
cycler==0.11.0
decorator==5.1.0
discogs-client==2.3.0
Django==4.1
django-environ==0.9.0
django-extensions==3.2.1
djangorestframework==3.14.0
docopt==0.6.2
environ==1.0
folium==0.12.1.post1
fonttools==4.28.1
graphene==2.1.9
graphene-django==2.15.0
graphql-core==2.3.2
graphql-relay==2.0.1
idna==3.3
ipython==7.29.0
jedi==0.18.1
Jinja2==3.0.3
joblib==1.1.0
kiwisolver==1.3.2
librosa==0.8.1
llvmlite==0.37.0
MarkupSafe==2.1.0
matplotlib==3.5.0
matplotlib-inline==0.1.3
musicbrainzngs==0.7.1
numba==0.54.1
numpy==1.20.3
oauthlib==3.2.2
packaging==21.2
pandas==1.5.3
parso==0.8.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.4.0
pip-tools==6.12.1
pipreqs==0.4.11
pooch==1.5.2
promise==2.3
prompt-toolkit==3.0.22
psycopg2==2.9.3
ptyprocess==0.7.0
pycparser==2.21
Pygments==2.10.0
pyparsing==2.4.7
pyproject_hooks==1.0.0
python-dateutil==2.8.2
python-decouple==3.6
pytz==2022.7.1
regex==2022.10.31
requests==2.26.0
resampy==0.2.2
Rx==1.6.1
scikit-learn==1.0.1
scipy==1.7.2
setuptools-scm==6.3.2
singledispatch==3.7.0
six==1.16.0
SoundFile==0.10.3.post1
sqlparse==0.4.2
text-unidecode==1.3
threadpoolctl==3.0.0
tomli==1.2.2
traitlets==5.1.1
urllib3==1.26.7
wcwidth==0.2.5
yarg==0.1.9
